change theta to double andd minor clean up

+++ ml_c/src/cost_function.c
@@ -17,7 +17,7 @@
 
 #include "cost_function.h"
 
-double cost_function(double* X, double* y, float theta, int m)
+double cost_function(double* X, double* y, double theta, int m)
 {
     /*
         Creating the algorithm for the cost function.
@@ -64,7 +64,7 @@ double cost_function(double* X, double* y, float theta, int m)
     return J_theta;
 }
 
-double cost_function_multiple(double* X, double* y, double *theta,
+double cost_function_multiple(double* X, double* y, double* theta,
                                                                 int m)
 {
     /*
